This commit transforms PostLink styled component into a link tag.
